Extract UDP transport in something reusable
